Don't reset authorization header if no auth data is stored (#265)
Habmin resets the Authorization header when clearing stale saved auth info, but does so even if no auth info was saved (e.g., when using nginx for authentication instead). This is the cause of #220.

This change only clears the auth header when stale saved data is actually being cleared as well.